Rev sym, trasselp, ny pdf, OddGen, ej bakofvp.

+++ Ovn-4 Minneshantering/Fibonacci_I.cs
@@ -10,10 +10,10 @@ namespace Ovn_4_Minneshantering // 0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144,
     {
         public static double Fibonaccisekvens(int n)
         {
-            int a = 0, b = 1, c;
+            ulong a = 0, b = 1, c; // Utökad gräns med ulong.
             if (n == 0) return a;
 
-            for (int i = 2; i <= n; i++)
+            for (long i = 2; i <= n; i++) // Utökad gräns med long.
             {
                 c = a + b;
                 a = b;
